天花板上的灯

+++ 96.天花板上的小灯.c
@@ -8,8 +8,8 @@
 #include <stdio.h>
 
 int main() {
-    int a, b;
-    scanf("%d%d", &a, &b);
-    printf("%d\n", a * b);
+    long a, b;
+    scanf("%ld%ld", &a, &b);
+    printf("%ld\n", a * b);
     return 0;
 }
